+++ tpl/fiche_of.tpl.php
@@ -27,7 +27,7 @@
 				</tr>
 				[onshow;block=begin;when [assetOf.id]=0]
 				<tr><td>[view.langs.transnoentities(ProductToProduce)]</td><td>[assetOf.product_to_create;strconv=no;protect=no]</td></tr>
-				<tr><td>Quantité à produire</td><td>[assetOf.quantity_to_create;strconv=no;protect=no]</td></tr>
+				<tr><td>[view.langs.transnoentities(QtyToMake)]</td><td>[assetOf.quantity_to_create;strconv=no;protect=no]</td></tr>
 				[onshow;block=end]
 				<tr><td>[view.langs.transnoentities(ParentOF)]</td><td>[assetOf.link_assetOf_parent;strconv=no;protect=no;magnet=tr]</td></tr>
 				<tr rel="fk_commande">